added color bar option to Slack module
This update will allow people to add a color bar at the front of a
Slack notification using the default 3 colors by name Slack specify
(good, warning, danger).

If no color is specified, or the default is used (normal) then no bar
will be added.

Description and example also added in this update.

Color bars are added by using the attachments json object inside the
payload - this is a very simplistic implementation as using custom
colors or adding titles or other formatting are not included in this
update and if needed I’m sure somebody else can spend the time to add
them later…

Tested with ansible 1.7