Reverse the unpack list operation
Instead of doing an unpack, deliberately specify which parameters you
want to use. This allows us to flexibly add more parameters to the
f5_argument_spec without having to rewrite all the modules that use
it.

Functionally this commit changes nothing, it just provides for a
different way of accessing the parameters to the module