openbsd_pkg: Use correct part of name in match. (#3151)
* openbsd_pkg: Use correct part of name in match.

Previously this part of the code could assume that the name was a stem
with nothing else attached (like "autoconf"). With the introduction of the
branch syntax ("autoconf%2.13") this is no longer true.

Check if the package name was identified as using a "branch" style name,
and base the match on the leading part of the name if that is the case.

While here remove unnecessary "pass" and tidy up debug log message.

Problem reported by @jasperla.

* openbsd_pkg: Add missing "." in comment.